Componentize the forms and update the webhook POST request
